# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Local AI Core",
    "description": "Local AI Core backend for AI Companion Project.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "Health & System"
        ],
        "summary": "Public Health Check",
        "description": "Public liveness and readiness probe.\nAndroid uses this to detect if the Windows host PC is online on the local network.\nDoes not require pairing credentials.",
        "operationId": "health_check_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/status": {
      "get": {
        "tags": [
          "Health & System"
        ],
        "summary": "Host System Status (Protected)",
        "description": "Detailed Windows host machine telemetry.\nRequires valid pairing token.",
        "operationId": "system_status_api_v1_system_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Verify Pairing Token",
        "description": "Test pairing credentials.\nReturns 200 with confirmation if the provided token matches the host secret.\nReturns 401 if missing, malformed, or mismatched.",
        "operationId": "verify_credentials_api_v1_auth_verify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "List Tasks",
        "description": "List all tasks owned by the authenticated profile with optional filtering.",
        "operationId": "list_tasks_api_v1_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter tasks by status.",
              "title": "Status"
            },
            "description": "Filter tasks by status."
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskPriority"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter tasks by priority.",
              "title": "Priority"
            },
            "description": "Filter tasks by priority."
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset for pagination.",
              "default": 0,
              "title": "Skip"
            },
            "description": "Offset for pagination."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page limit.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page limit."
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create Task",
        "description": "Create a new action item or reminder strictly scoped to the authenticated owner.",
        "operationId": "create_task_api_v1_tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get Task by ID",
        "description": "Retrieve a single task by its unique UUID identifier.",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Tasks"
        ],
        "summary": "Partial Update Task",
        "description": "Partially update specific fields on an existing task.",
        "operationId": "update_task_api_v1_tasks__task_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Delete Task",
        "description": "Permanently delete a task by ID.",
        "operationId": "delete_task_api_v1_tasks__task_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthVerifyResponse": {
        "properties": {
          "authenticated": {
            "type": "boolean",
            "title": "Authenticated",
            "description": "Indicates token passed constant-time verification.",
            "default": true
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Expected authorization scheme.",
            "default": "Bearer"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Status confirmation message.",
            "default": "Token is valid and authenticated."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AuthVerifyResponse",
        "description": "Result returned on token verification check."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Overall system health status.",
            "default": "healthy"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Application version."
          },
          "database_connected": {
            "type": "boolean",
            "title": "Database Connected",
            "description": "SQLite database read/write connectivity flag."
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "UTC server timestamp."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "database_connected",
          "timestamp"
        ],
        "title": "HealthResponse",
        "description": "Liveness and dependency connectivity status."
      },
      "SystemStatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "online"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "description": "Operating system identifier (e.g. Windows 11)."
          },
          "python_version": {
            "type": "string",
            "title": "Python Version",
            "description": "Python interpreter version."
          },
          "hostname": {
            "type": "string",
            "title": "Hostname",
            "description": "Local machine hostname."
          },
          "cpu_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpu Count",
            "description": "Logical CPU core count."
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "UTC server timestamp."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "platform",
          "python_version",
          "hostname",
          "timestamp"
        ],
        "title": "SystemStatusResponse",
        "description": "Host machine hardware profile and runtime status."
      },
      "TaskCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "description": "Short title of the task."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Detailed instructions or context."
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority",
            "description": "Urgency rating.",
            "default": "medium"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date",
            "description": "Target completion timestamp in ISO-8601."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TaskCreate",
        "description": "Payload to create a new task."
      },
      "TaskListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "List of matching tasks."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total count of tasks matching query."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total"
        ],
        "title": "TaskListResponse",
        "description": "Paginated or listed collection of tasks."
      },
      "TaskPriority": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "urgent"
        ],
        "title": "TaskPriority",
        "description": "Urgency ratings for scheduling."
      },
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique UUID identifier."
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "description": "Account/profile ownership identifier."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Task title."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Task notes."
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "description": "Current lifecycle state."
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority",
            "description": "Urgency rating."
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date",
            "description": "Due date timestamp."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp in UTC."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last modification timestamp in UTC."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "title",
          "status",
          "priority",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse",
        "description": "Output representation of a stored task."
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "pending",
          "in_progress",
          "completed",
          "cancelled"
        ],
        "title": "TaskStatus",
        "description": "Lifecycle statuses for companion tasks and action items."
      },
      "TaskUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskPriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TaskUpdate",
        "description": "Payload to update an existing task. Fields omitted remain unchanged."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}